Add all the crud operations to each entity

+++ V2/Server/MedicalThyroidReports/Modals/Medecin.cs
@@ -5,10 +5,10 @@ namespace MedicalThyroidReports.Modals
     public class Medecin
     {
         [Key] 
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Prenom { get; set; }
         public string Nom { get; set; }
-        public long IdSpecialite { get; set; }
+        public int IdSpecialite { get; set; }
         public string Grade { get; set; }
     }
 }
